+++ src/main/java/com/kakaoseventeen/dogwalking/chat/domain/ChatMessage.java
@@ -1,4 +1,4 @@
-package com.kakaoseventeen.dogwalking.chat.model;
+package com.kakaoseventeen.dogwalking.chat.domain;
 
 import com.kakaoseventeen.dogwalking.member.domain.Member;
 import jakarta.persistence.*;
