+++ src/main/java/com/kakaoseventeen/dogwalking/member/domain/Member.java
@@ -15,7 +15,7 @@ import java.math.BigDecimal;
 @Table(name="member_tb")
 public class Member {
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @Column(length = 45, nullable = false)
     private String nickname;
