+++ src/main/java/com/kakaoseventeen/dogwalking/walkRoad/domain/WalkRoad.java
@@ -22,7 +22,7 @@ import java.time.LocalDateTime;
 public class WalkRoad {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
