+++ src/test/java/com/kakaoseventeen/dogwalking/payment/PaymentRespTest.java
@@ -1,10 +1,8 @@
 package com.kakaoseventeen.dogwalking.payment;
 
-import com.kakaoseventeen.dogwalking.payment.domain.Payment;
 import com.kakaoseventeen.dogwalking.payment.repository.PaymentRepository;
-import com.kakaoseventeen.dogwalking.walk.domain.Walk;
 import com.kakaoseventeen.dogwalking.walk.repository.WalkRepository;
-import com.kakaoseventeen.dogwalking.chat.model.ChatRoom;
+import com.kakaoseventeen.dogwalking.chat.domain.ChatRoom;
 import com.kakaoseventeen.dogwalking.chat.repository.ChatRoomRepository;
 import com.kakaoseventeen.dogwalking.dog.domain.Dog;
 import com.kakaoseventeen.dogwalking.dog.repository.DogJpaRepository;
