+++ src/test/java/com/kakaoseventeen/dogwalking/walkRoad/WalkRoadControllerTest.java
@@ -2,12 +2,10 @@ package com.kakaoseventeen.dogwalking.walkRoad;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.kakaoseventeen.dogwalking._core.utils.GetEntity;
-import com.kakaoseventeen.dogwalking.walk.domain.Walk;
 import com.kakaoseventeen.dogwalking.walk.repository.WalkRepository;
-import com.kakaoseventeen.dogwalking.walkRoad.domain.WalkRoad;
 import com.kakaoseventeen.dogwalking.walkRoad.dto.WalkRoadReqDTO;
 import com.kakaoseventeen.dogwalking.walkRoad.repository.WalkRoadRepository;
-import com.kakaoseventeen.dogwalking.chat.model.ChatRoom;
+import com.kakaoseventeen.dogwalking.chat.domain.ChatRoom;
 import com.kakaoseventeen.dogwalking.chat.repository.ChatRoomRepository;
 import com.kakaoseventeen.dogwalking.member.domain.Member;
 import com.kakaoseventeen.dogwalking.member.repository.MemberJpaRepository;
